# docs/bounded-contexts/catalog/api.yaml
openapi: 3.0.3
info:
  title: 商品限定上下文 API
  description: 类别、商品（SPU）、规格维度与选项、SKU 的 REST 契约。Step Definitions 与实现均按此规范。
  version: 0.1.0
paths:
...
  /api/products/{spuId}/dimensions:
    post:
      summary: 为 SPU 添加规格维度
      parameters:
        - name: spuId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpecDimensionCreate'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpecDimension'
        '404':
          description: SPU 不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '400':
          description: 业务校验失败（如同 SPU 内维度名称重复）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      summary: 查询某 SPU 下所有维度及选项
      parameters:
        - name: spuId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 维度列表（含各维度下选项）
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SpecDimensionWithOptions'
        '404':
          description: SPU 不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/products/{spuId}/dimensions/{dimensionId}/options:
    post:
      summary: 为某维度添加可选项
      parameters:
        - name: spuId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: dimensionId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpecOptionCreate'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpecOption'
        '404':
          description: SPU 或维度不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '400':
          description: 业务校验失败（如同维度内选项值重复）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ---------- 产品级展示图 ----------
...
  /api/products/{spuId}/dimensions/{dimensionId}/options/{optionId}/images:
    post:
      summary: 为选项添加展示图
      description: 任意维度的选项均可添加展示图，不再要求「影响外观」。
      parameters:
        - name: spuId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: dimensionId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: optionId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OptionImageCreate'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OptionImage'
        '404':
          description: SPU、维度或选项不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '400':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      summary: 查询某选项下所有展示图
      parameters:
        - name: spuId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: dimensionId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: optionId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 展示图列表（按 sortOrder 排序）
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OptionImage'
        '404':
          description: SPU、维度或选项不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/products/{spuId}/dimensions/{dimensionId}/options/{optionId}/images/{imageId}:
    delete:
      summary: 删除展示图
      parameters:
        - name: spuId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: dimensionId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: optionId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: imageId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: 删除成功
        '404':
          description: 展示图不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ---------- SKU ----------
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        message:
          type: string
          description: 错误说明，便于 Step 断言

    # ---------- 规格维度与选项 ----------
    SpecDimension:
      type: object
      required: [id, spuId, name, required]
      properties:
        id:
          type: integer
          format: int64
        spuId:
          type: integer
          format: int64
        name:
          type: string
        required:
          type: boolean
          description: 创建 SKU 时是否必选
        sortOrder:
          type: integer
          nullable: true
    SpecDimensionCreate:
      type: object
      required: [name, required]
      properties:
        name:
          type: string
        required:
          type: boolean
        sortOrder:
          type: integer
          nullable: true
    SpecOption:
      type: object
      required: [id, specDimensionId, optionValue]
      properties:
        id:
          type: integer
          format: int64
        specDimensionId:
          type: integer
          format: int64
        optionValue:
          type: string
        sortOrder:
          type: integer
          nullable: true
        images:
          type: array
          description: 该选项的展示图列表，按 sortOrder 排序
          items:
            $ref: '#/components/schemas/OptionImage'
    SpecOptionCreate:
      type: object
      required: [optionValue]
      properties:
        optionValue:
          type: string
        sortOrder:
          type: integer
          nullable: true
    SpecDimensionWithOptions:
      type: object
      description: 维度 + 其下选项列表（GET 维度及选项时用）
      allOf:
        - $ref: '#/components/schemas/SpecDimension'
        - type: object
          properties:
            options:
              type: array
              items:
                $ref: '#/components/schemas/SpecOption'
    OptionImage:
      type: object
      required: [id, specOptionId, imageUrl]
      properties:
        id:
          type: integer
          format: int64
        specOptionId:
          type: integer
          format: int64
        imageUrl:
          type: string
        sortOrder:
          type: integer
          nullable: true
    OptionImageCreate:
      type: object
      required: [imageUrl]
      properties:
        imageUrl:
          type: string
        sortOrder:
          type: integer
          nullable: true
...